Move code to pkg
