+++ pkg/ingress/kube/annotations/auth_test.go
@@ -29,7 +29,7 @@ import (
 	listerv1 "k8s.io/client-go/listers/core/v1"
 	"k8s.io/client-go/tools/cache"
 
-	"github.com/alibaba/higress/ingress/kube/util"
+	"github.com/alibaba/higress/pkg/ingress/kube/util"
 )
 
 func TestAuthParse(t *testing.T) {
