+++ pkg/ingress/kube/annotations/cors.go
