+++ pkg/ingress/kube/annotations/header_control.go
@@ -19,7 +19,7 @@ import (
 
 	networking "istio.io/api/networking/v1alpha3"
 
-	. "github.com/alibaba/higress/ingress/log"
+	. "github.com/alibaba/higress/pkg/ingress/log"
 )
 
 const (
