+++ pkg/ingress/kube/annotations/upstreamtls.go
@@ -22,7 +22,7 @@ import (
 	networking "istio.io/api/networking/v1alpha3"
 	"istio.io/istio/pilot/pkg/model/credentials"
 
-	"github.com/alibaba/higress/ingress/kube/util"
+	"github.com/alibaba/higress/pkg/ingress/kube/util"
 )
 
 const (
