+++ pkg/ingress/kube/ingress/status.go
@@ -12,10 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package ingressv1
+package ingress
 
 import (
 	"context"
+	common2 "github.com/alibaba/higress/pkg/ingress/kube/common"
+	. "github.com/alibaba/higress/pkg/ingress/log"
 	"reflect"
 	"sort"
 	"time"
@@ -26,11 +28,8 @@ import (
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/client-go/kubernetes"
 	listerv1 "k8s.io/client-go/listers/core/v1"
-	ingresslister "k8s.io/client-go/listers/networking/v1"
+	ingresslister "k8s.io/client-go/listers/networking/v1beta1"
 	"k8s.io/client-go/tools/cache"
-
-	"github.com/alibaba/higress/ingress/kube/common"
-	. "github.com/alibaba/higress/ingress/log"
 )
 
 // statusSyncer keeps the status IP in each Ingress resource updated
@@ -52,8 +51,8 @@ func newStatusSyncer(localKubeClient, client kubelib.Client, controller *control
 		client:             client,
 		controller:         controller,
 		watchedNamespace:   namespace,
-		ingressLister:      client.KubeInformer().Networking().V1().Ingresses().Lister(),
-		ingressClassLister: client.KubeInformer().Networking().V1().IngressClasses().Lister(),
+		ingressLister:      client.KubeInformer().Networking().V1beta1().Ingresses().Lister(),
+		ingressClassLister: client.KubeInformer().Networking().V1beta1().IngressClasses().Lister(),
 		// search service in the mse vpc
 		serviceLister: localKubeClient.KubeInformer().Core().V1().Services().Lister(),
 	}
@@ -62,7 +61,7 @@ func newStatusSyncer(localKubeClient, client kubelib.Client, controller *control
 func (s *statusSyncer) run(stopCh <-chan struct{}) {
 	cache.WaitForCacheSync(stopCh, s.controller.HasSynced)
 
-	ticker := time.NewTicker(common.DefaultStatusUpdateInterval)
+	ticker := time.NewTicker(common2.DefaultStatusUpdateInterval)
 	for {
 		select {
 		case <-stopCh:
@@ -77,14 +76,14 @@ func (s *statusSyncer) run(stopCh <-chan struct{}) {
 }
 
 func (s *statusSyncer) runUpdateStatus() error {
-	svcList, err := s.serviceLister.Services(s.watchedNamespace).List(common.SvcLabelSelector)
+	svcList, err := s.serviceLister.Services(s.watchedNamespace).List(common2.SvcLabelSelector)
 	if err != nil {
 		return err
 	}
 
 	IngressLog.Debugf("found number %d of svc", len(svcList))
 
-	lbStatusList := common.GetLbStatusList(svcList)
+	lbStatusList := common2.GetLbStatusList(svcList)
 	if len(lbStatusList) == 0 {
 		return nil
 	}
@@ -111,7 +110,7 @@ func (s *statusSyncer) updateStatus(status []coreV1.LoadBalancerIngress) error {
 		}
 
 		curIPs := ingress.Status.LoadBalancer.Ingress
-		sort.SliceStable(curIPs, common.SortLbIngressList(curIPs))
+		sort.SliceStable(curIPs, common2.SortLbIngressList(curIPs))
 
 		if reflect.DeepEqual(status, curIPs) {
 			IngressLog.Debugf("skipping update of Ingress %v/%v within cluster %s (no change)",
@@ -122,7 +121,7 @@ func (s *statusSyncer) updateStatus(status []coreV1.LoadBalancerIngress) error {
 		ingress.Status.LoadBalancer.Ingress = status
 		IngressLog.Infof("Update Ingress %v/%v within cluster %s status",
 			ingress.Namespace, ingress.Name, s.controller.options.ClusterId)
-		_, err = s.client.NetworkingV1().Ingresses(ingress.Namespace).UpdateStatus(context.TODO(), ingress, metaV1.UpdateOptions{})
+		_, err = s.client.NetworkingV1beta1().Ingresses(ingress.Namespace).UpdateStatus(context.TODO(), ingress, metaV1.UpdateOptions{})
 		if err != nil {
 			IngressLog.Warnf("error updating ingress %s/%s within cluster %s status: %v",
 				ingress.Namespace, ingress.Name, s.controller.options.ClusterId, err)
