+++ pkg/ingress/mcp/generator.go
